Match column format on dlog_15 and dlog_90 to other similar-style views.

+++ fannie/install/sql/trans/dtransactions.php
@@ -228,7 +228,7 @@ $CREATE['trans.dtransactions'] = "
 	  `matched` smallint(6) default NULL,
 	  `memType` tinyint(2) default NULL,
 	  `staff` tinyint(4) default NULL,
-	  `numflag` smallint(6) default 0 NULL,
+	  `numflag` int(11) default 0 NULL,
 	  `charflag` varchar(2) default '' NULL,
 	  `card_no` int(11) default NULL,
 	  `trans_id` int(11) default NULL,
